refactor(transform): isolate static safety proofs
